[COST-3797] Remove Depriecated Settings.
