+++ .unleash/flags.json
@@ -2,22 +2,6 @@
     "version": 1,
     "features": [
         {
-            "name": "cost-management.backend.disable-ddf-tag-form",
-            "description": "Toggle to disabled tag data driven form for settings.",
-            "type": "killswitch",
-            "project": "default",
-            "enabled": false,
-            "stale": false,
-            "strategies": [
-                {
-                    "name": "default",
-                    "constraints": []
-                }
-            ],
-            "variants": [],
-            "createdAt": "2023-07-25T11:50:00.756Z"
-        },
-        {
             "name": "cost-management.backend.disable-source",
             "description": "Toggle to disable processing for a source",
             "type": "permission",
@@ -241,25 +225,6 @@
             ],
             "variants": [],
             "createdAt": "2022-09-08T09:48:54.492Z"
-        },
-        {
-            "name": "cost-management.backend.enable_aws_category_settings",
-            "description": "any account listed in this strategy will be allowed to enable AWS Category settings.",
-            "type": "permission",
-            "project": "default",
-            "enabled": true,
-            "stale": false,
-            "strategies": [
-                {
-                    "name": "schema-strategy",
-                    "parameters": {
-                        "schema-name": ""
-                    },
-                    "constraints": []
-                }
-            ],
-            "variants": [],
-            "createdAt": "2023-04-06T09:40:54.492Z"
         }
     ],
     "strategies": [
