--- koku/api/settings/__init__.py
@@ -1,1 +1,0 @@
-# noqa
