+++ koku/api/urls.py
@@ -87,7 +87,6 @@ from api.views import ResourceTypeView
 from api.views import SettingsAWSCategoryKeyView
 from api.views import SettingsDisableAWSCategoryKeyView
 from api.views import SettingsEnableAWSCategoryKeyView
-from api.views import SettingsView
 from api.views import StatusView
 from api.views import UserAccessView
 from api.views import UserCostTypeSettings
@@ -342,7 +341,6 @@ urlpatterns = [
     ),
     path("ingress/reports/", IngressReportsView.as_view(), name="reports"),
     path("ingress/reports/<source>/", IngressReportsDetailView.as_view(), name="reports-detail"),
-    path("settings/", SettingsView.as_view(), name="settings"),
     path("settings/aws_category_keys/", SettingsAWSCategoryKeyView.as_view(), name="settings-aws-category-keys"),
     path(
         "settings/aws_category_keys/enable/",
