+++ koku/api/user_settings/aws_category_keys/view.py
@@ -16,8 +16,8 @@ from api.common.pagination import ListPaginator
 from api.common.permissions.settings_access import SettingsAccessPermission
 from api.provider.models import Provider
 from api.report.constants import URL_ENCODED_SAFE
-from api.settings.aws_category_keys.serializers import SettingsAWSCategoryKeyIDSerializer
-from api.settings.aws_category_keys.serializers import SettingsAWSCategoryKeySerializer
+from api.user_settings.aws_category_keys.serializers import SettingsAWSCategoryKeyIDSerializer
+from api.user_settings.aws_category_keys.serializers import SettingsAWSCategoryKeySerializer
 from koku.cache import invalidate_view_cache_for_tenant_and_source_type
 from reporting.provider.aws.models import AWSEnabledCategoryKeys
