+++ koku/api/user_settings/views.py
@@ -20,12 +20,12 @@ from rest_framework.views import APIView
 from api.common.pagination import ListPaginator
 from api.common.permissions.settings_access import SettingsAccessPermission
 from api.provider.models import Provider
-from api.settings.utils import set_cost_type
-from api.settings.utils import set_currency
 from api.user_settings.serializer import UserSettingSerializer
 from api.user_settings.serializer import UserSettingUpdateCostTypeSerializer
 from api.user_settings.serializer import UserSettingUpdateCurrencySerializer
 from api.user_settings.settings import COST_TYPES
+from api.user_settings.utils import set_cost_type
+from api.user_settings.utils import set_currency
 from api.utils import get_account_settings
 from api.utils import get_cost_type
 from api.utils import get_currency
