+++ koku/masu/test/util/test_common.py
@@ -439,88 +439,6 @@ class CommonUtilTests(MasuTestCase):
         self.assertEqual(enabled, check_enabled)
         self.assertEqual(orig_disabled, check_disabled)
 
-    def test_update_enabled_keys_aws(self):
-        with schema_context(self.schema):
-            orig_keys = [
-                {"key": e.key, "enabled": e.enabled, "provider_type": "AWS"}
-                for e in EnabledTagKeys.objects.filter(provider_type=Provider.PROVIDER_AWS)
-            ]
-            EnabledTagKeys.objects.filter(provider_type=Provider.PROVIDER_AWS).delete()
-            for key in ("masu", "database", "processor", "common"):
-                EnabledTagKeys.objects.create(key=key, enabled=(key != "masu"), provider_type=Provider.PROVIDER_AWS)
-            all_keys = list(EnabledTagKeys.objects.filter(provider_type=Provider.PROVIDER_AWS))
-
-        orig_disabled = {e.key for e in all_keys if not e.enabled}
-        orig_enabled = {e.key for e in all_keys if e.enabled}
-        disabled = None
-        enabled = set()
-        for i, k in enumerate(orig_enabled):
-            if i == 0:
-                disabled = k
-            else:
-                enabled.add(k)
-        new_keys = {"ek_test1", "ek_test2"}
-        enabled.update(new_keys)
-
-        common_utils.update_enabled_keys(self.schema, EnabledTagKeys, enabled)
-        with schema_context(self.schema):
-            all_keys = list(EnabledTagKeys.objects.filter(provider_type=Provider.PROVIDER_AWS))
-            EnabledTagKeys.objects.filter(provider_type=Provider.PROVIDER_AWS).delete()
-            EnabledTagKeys.objects.bulk_create([EnabledTagKeys(**rec) for rec in orig_keys])
-
-        all_keys_set = {k.key for k in all_keys}
-        check_disabled = {d.key for d in all_keys if not d.enabled}
-        check_enabled = {e.key for e in all_keys if e.enabled}
-
-        self.assertTrue(new_keys.isdisjoint(all_keys))
-        self.assertTrue(disabled in all_keys_set)
-        self.assertTrue(disabled in check_disabled)
-        self.assertEqual(orig_disabled.intersection(check_disabled), orig_disabled)
-        self.assertNotEqual(orig_disabled, check_disabled)
-        self.assertNotEqual(orig_enabled, check_enabled)
-        self.assertEqual((enabled - new_keys), check_enabled)
-
-    def test_update_enabled_keys_azure(self):
-        with schema_context(self.schema):
-            orig_keys = [
-                {"key": e.key, "enabled": e.enabled, "provider_type": "Azure"}
-                for e in EnabledTagKeys.objects.filter(provider_type=Provider.PROVIDER_AZURE)
-            ]
-            EnabledTagKeys.objects.filter(provider_type=Provider.PROVIDER_AZURE).delete()
-            for key in ("masu", "database", "processor", "common"):
-                EnabledTagKeys.objects.create(key=key, enabled=(key != "masu"), provider_type=Provider.PROVIDER_AZURE)
-            all_keys = list(EnabledTagKeys.objects.filter(provider_type=Provider.PROVIDER_AZURE))
-
-        orig_disabled = {e.key for e in all_keys if not e.enabled}
-        orig_enabled = {e.key for e in all_keys if e.enabled}
-        disabled = None
-        enabled = set()
-        for i, k in enumerate(orig_enabled):
-            if i == 0:
-                disabled = k
-            else:
-                enabled.add(k)
-        new_keys = {"ek_test1", "ek_test2"}
-        enabled.update(new_keys)
-
-        common_utils.update_enabled_keys(self.schema, EnabledTagKeys, enabled)
-        with schema_context(self.schema):
-            all_keys = list(EnabledTagKeys.objects.filter(provider_type=Provider.PROVIDER_AZURE))
-            EnabledTagKeys.objects.filter(provider_type=Provider.PROVIDER_AZURE).delete()
-            EnabledTagKeys.objects.bulk_create([EnabledTagKeys(**rec) for rec in orig_keys])
-
-        all_keys_set = {k.key for k in all_keys}
-        check_disabled = {d.key for d in all_keys if not d.enabled}
-        check_enabled = {e.key for e in all_keys if e.enabled}
-
-        self.assertTrue(new_keys.isdisjoint(all_keys))
-        self.assertTrue(disabled in all_keys_set)
-        self.assertTrue(disabled in check_disabled)
-        self.assertEqual(orig_disabled.intersection(check_disabled), orig_disabled)
-        self.assertNotEqual(orig_disabled, check_disabled)
-        self.assertNotEqual(orig_enabled, check_enabled)
-        self.assertEqual((enabled - new_keys), check_enabled)
-
     def test_strip_characters_from_column_name(self):
         """Test that column names are converted properly."""
         bad_str = r"column\one:two-three four,five/six_seven"
